requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
WeasyPrint==60.1
pytest
pip-audit
